// app/schemas/file.ts
export interface User {
    id?: number;
    name: string;
    email: string;
    created_at?: string;
  }


export interface event_registration {
    id?: number;  // Optional ID, auto-generated by the database
    time_slot: string;  // Time slot for the event
    course_programs?: number[];  // Optional array of course programs
    lecture_preference: string;  // Preferred lecture type
    parent_title?: string;  // Optional parent title
    parent_first_name: string;  // Parent's first name
    parent_last_name: string;  // Parent's last name
    parent_email: string;  // Parent's email address
    parent_phone: string;  // Parent's phone number
    student_title?: string;  // Optional student title
    student_first_name: string;  // Student's first name
    student_last_name: string;  // Student's last name
    current_school: string;  // Student's current school
    grade_level: string;  // Grade level of the student
    total_participants: number;  // Total number of participants (greater than 0)
    created_at?: string;  // Optional created_at timestamp
}
  

export interface role {
  id?: number;
  user_id: string | null;
  email: string | null;
  role: string;
  form: boolean;
  base1: boolean | null;
  base2: boolean | null;
  base3: boolean | null;
  base4: boolean | null;
  base5: boolean | null;
  base6: boolean | null;
  base7: boolean | null;
  base8: boolean | null;
  base9: boolean | null;
  base10: boolean | null;
  base11: boolean | null;
  base12: boolean | null;
}